Fix trivial compilation errors
Fixed uninitiated variables, implicit type casts, C++ includes, ambiguous methods.

+++ Code/BandTest/BandTest.cpp
@@ -784,7 +784,7 @@ unsigned long Upstream_Detect(unsigned long server_ip, unsigned long my_ip, int 
 	*/
 	unsigned long downstream_bandwidth = upstream_bandwidth;
 	int old_band = Get_Registry_Int("Up", 0);
-	unsigned long diff = abs(upstream_bandwidth - old_band);
+	unsigned long diff = abs((long)(upstream_bandwidth - old_band));
 	bool calc_down = true;
 	if (diff < upstream_bandwidth / 10) {
 		downstream_bandwidth = Get_Registry_Int("Down", upstream_bandwidth);
